+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Release History
 
+## 1.29.1 (2025-03-12)
+
+### Snowpark Python API Updates
+
+#### Bug Fixes
+
+- Fixed a bug in `DataFrameReader.dbapi` (PrPr) that prevents usage in stored procedure and snowbooks.
+
 ## 1.29.0 (2025-03-05)
 
 ### Snowpark Python API Updates